rename: AbstractEntity.kt 패키지 변경
